Move low-level apt stuff to utils/packages.py
